chore: Use editable installs for cloned MCs in development mode
This uses editable installs (`pip install -e ...`) for MCs that have
been cloned via the `--clone` option to the install script. Otherwise,
the online/hosted PyPI versions are used. To make the conditional more
clear, this also adds a local variable to name the function argument as
well as reorders the conditional to explicitly use the `dev` optional
dependency by name ahead of its reference in the comments.