feat: Setting configuration with JSON string (#166)
This PR adds the ability for users to take in a properly formatted JSON
string and set the FIREWHEEL configuration.

This JSON string will set/replace a _subset_ of the configuration and
any keys or sub-keys (or sub-sub-keys) not present **will** not be
impacted or overwritten.
The string _must_ include the top-level config key as well as any
sub-keys needing modifications.

For example, to change the value for the config key ``logging.level``
from DEBUG to INFO, here is how that would not look:

**Before Config**
```yaml
logging:
    cli_log: cli.log
    discovery_log: discovery.log
    firewheel_log: firewheel.log
    level: DEBUG
    minimega_log: minimega.log
    root_dir: /scratch/firewheel
    vmr_log_dir: vm_resource_logs
```

```bash
firewheel config set -j '{"logging":{"level":"INFO"}}'
```

**After Config**
```yaml
logging:
    cli_log: cli.log
    discovery_log: discovery.log
    firewheel_log: firewheel.log
    level: INFO
    minimega_log: minimega.log
    root_dir: /scratch/firewheel
    vmr_log_dir: vm_resource_logs
```

---------

Signed-off-by: Steven Elliott <6461548+sdelliot@users.noreply.github.com>
Co-authored-by: Mitch Negus <21086604+mitchnegus@users.noreply.github.com>